pyproject.toml
[project]
name = "response_evaluator"
version = "0.0.0"
description = "A generic prompt-based validator for Guardrails AI"
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0",
    "litellm>=1.23.0"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]